refactor(oracle): migrate ipsec generator samplers with post-sample crypto hook
